# DOC/output/runs/20260502-024248/specs/openapi.yaml
openapi: 3.1.0
info:
  title: AU Plumber Website API
  version: 1.0.0
  description: Internal API for the Australian plumbing lead-generation website.
servers:
  - url: https://yourdomain.com.au
    description: Production
  - url: http://localhost:3000
    description: Local development
...
paths:
...
  /api/webhooks/sanity:
    post:
      tags: [webhooks]
      summary: Sanity CMS revalidation webhook
      description: Triggered by Sanity on document publish/update/delete. Verifies signature and calls revalidatePath.
      responses:
        '200':
          description: Revalidation queued
        '400':
          description: Signature verification failed
...
  /api/draft:
    get:
      tags: [system]
      summary: Enable Sanity draft mode
      security:
        - clerkSession: []
      parameters:
        - name: secret
          in: query
          required: true
          schema:
            type: string
        - name: slug
          in: query
          schema:
            type: string
      responses:
        '307':
          description: Redirect to draft page
...
components:
  securitySchemes:
    clerkSession:
      type: apiKey
      in: cookie
      name: __session
      description: Clerk session cookie
...